Move common styles for sectoral circumstnances components to one stylesheet and remove the duplicates
