+++ app/javascript/app/pages/regions/sectoral-circumstances/energy/energy-component.jsx
@@ -6,7 +6,7 @@ import Chart from 'components/chart';
 import { format } from 'd3-format';
 import CustomTooltip from '../bar-chart-tooltip';
 
-import styles from './energy-styles.scss';
+import styles from '../sectoral-circumstances-styles.scss';
 
 class Energy extends PureComponent {
   render() {
