+++ app/javascript/app/pages/regions/sectoral-circumstances/sectoral-circumstances-styles.scss
@@ -6,3 +6,27 @@
 
   margin-top: $gutter-padding;
 }
+
+.title {
+  font-weight: bold;
+  font-size: $font-size-large;
+  margin-bottom: 30px;
+}
+
+.cardHeader {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.cardWrapper {
+  margin-bottom: 50px;
+}
+
+.buttonWrapper {
+  background-color: $white;
+}
+
+.legend {
+  margin-top: 20px;
+}
